+++ lib/strong_migrations/adapters/mariadb_adapter.rb
@@ -6,7 +6,7 @@ module StrongMigrations
       end
 
       def min_version
-        "10.1"
+        "10.2"
       end
 
       def server_version
@@ -19,10 +19,6 @@ module StrongMigrations
 
       def set_statement_timeout(timeout)
         select_all("SET max_statement_time = #{connection.quote(timeout)}")
-      end
-
-      def online_ddl_supported?
-        server_version >= Gem::Version.new("10.0")
       end
 
       def add_column_default_safe?